#68 Fix autosave not enough space issue.
Problem was with using AutoSave filename for saving instead of active file name. This was causing during default autosave logic since it is relying on the filename to contain special text pattern (`Cycle XX`), while autosave file does not have it.
Fixed by asung current filename (which is in the right format).