#92 Add general event and command for side screens.
Currently it filters out all side screen since they were not tested.
Idea is to go through each of them and make sure that the event and command covers all required action syncs. It is possible that some sidescreens couldn't be easily synced through general event. However the idea is to cover 80%+ of sidescreens with general to avoid creating 70+ different events and commands.