#13 Introduce ChoreStateEvents.cs (#311)
#13 Introduce ChoreStateEvents.cs

The event is part of chore stat sync. This sync will be used for dynamic chores to properly sync their argument changes upon state transition. It might be not enough to cover all cases of parameter changes but this will cover some of cases for sure.
The event already enabled for some chores (3 of them) where this approach will work for sure.