Add chore driver synchronization
It adds a chores queue to a client`s driver and populates it with supported chores.
When server`s driver chooses a chore from supported to unsupported client`s driver
queue will be reset and repopulated when supported chore is chosen.

There will be conflicts until all chores are synchronized, for example with
drag tools, because they create chores that can be blocked by synced ones on
a client.

It's unknown if it's a temporary solution, probably it is, because eventually
a client won't create unsupported chores by itself.